Refactor NotesFilter to separate container
