+++ src/components/App/index.js
@@ -9,8 +9,8 @@ export const App = () => {
     <main className='a-404'>
       <h2>404 &ndash; Path not found.</h2>
       <p>No routes match <code>{location.pathname}</code>.</p>
-      <p>Please click the the 'Take Note' or 'Notable Notes' buttons above or your browser's forward and back buttons to navigate to your desired destination.</p>
-      <p>The Nordstrom Notes logo will return you to the Take Note page.</p>
+      <p>Please click the the 'Make Note' or 'Notable Notes' buttons above or your browser's forward and back buttons to navigate to your desired destination.</p>
+      <p>The Nordstrom Notes logo will return you to the Make Note page.</p>
     </main>
   )
