+++ src/containers/NotesList/index.js
@@ -1,6 +1,8 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { PropTypes } from 'prop-types';
+import { caughtError } from '../../actions';
+import NotesFilter from '../NotesFilter';
 import './NotesList.css';
 
 export class NotesList extends Component {
@@ -8,92 +10,25 @@ export class NotesList extends Component {
     super()
     this.state = {
       notes: [],
-      filter: '',
-      filterParams: '',
-      isDisabled: true,
-      placeholderText: 'Choose a filter.',
     }
   }
 
-  handleInputChange = async event => {
-    const { name, value } = event.target;
-
-    await this.setState({ [name]: value })
-    if (name === 'filter') this.toggleParams()
-  }
-
-  toggleParams = async () => {
-    const { filter } = this.state
-
-    filter
-      ? await this.setState({ isDisabled: false })
-      : await this.setState({ isDisabled: true })
-
-    this.formatPlaceholder(filter)
-  }
-
-  formatPlaceholder = async filter => {
-    const dateState = { placeholderText: 'Ex: 12-18-2018' }
-    const tagState = { placeholderText: 'Ex: Work' }
-
-    filter === 'date'
-      ? await this.setState(dateState)
-      : await this.setState(tagState)
-  }
+  noNotes = () => (
+    <section className='nl-list'>
+      <h3>You have no notes.</h3>
+      <p>Please click the the 'Make Note' or your browser's back button to navigate to the Make Note form.</p>
+      <p>The Nordstrom Notes logo will also return you to the Make Note page.</p>
+    </section>
+  )
 
   render() {
-    const { notes, filter, filterParams, isDisabled, placeholderText } = this.state;
+    const { notes } = this.state;
 
     return (
       <section className='NotesList'>
         <h2>Notable Notes</h2>
-        <form className='nl-form'>
-          <div className='nl-filter-by'>
-            <label htmlFor='filter-group'>Filter by:</label>
-            <div
-              className="nl-filter-group"
-              id='filter-group'
-              name='filter-type'
-            >
-              <div className='nl-filter-type'>
-                <input
-                  type='radio'
-                  name='filter'
-                  value='date'
-                  id='date'
-                  onChange={this.handleInputChange}
-                />
-                <label htmlFor="date">Date</label>
-              </div>
-              <div className='nl-filter-type'>
-                <input
-                  type='radio'
-                  name='filter'
-                  value='tag'
-                  id='tag'
-                  onChange={this.handleInputChange}
-                />
-                <label htmlFor='tag'>Tag</label>
-              </div>
-            </div>
-          </div>
-          <div className='nl-filter-params-group'>
-            <label htmlFor='filter-params'>Search:</label>
-            <input
-              className='nl-filter-params'
-              id='filter-params'
-              name='filterParams'
-              value={filterParams}
-              maxLength={ filter === 'date' ? '10' : '8' }
-              onChange={this.handleInputChange}
-              disabled={isDisabled}
-              placeholder={placeholderText}
-            />
-          </div>
-        </form>
-        <section className='nl-list'>
-          { /* add table here */ }
-        </section>
+        <NotesFilter />
+        {!notes.length ? this.noNotes() : ''}
       </section>
     )
   }
@@ -103,8 +38,13 @@ export const mapStateToProps = (state) => ({
   notes: state.notes
 })
 
+export const mapDispatchToProps = (dispatch) => ({
+  caughtError: (errorMessage) => dispatch(caughtError(errorMessage)),
+})
+
 NotesList.propTypes = {
-  notes: PropTypes.array
+  notes: PropTypes.array,
+  caughtError: PropTypes.func,
 }
 
-export default connect(mapStateToProps)(NotesList)
+export default connect(mapStateToProps, mapDispatchToProps)(NotesList)
